Migrate all financial calculations from double/float to BigDecimal
XOF currency amounts use zero decimal places, percentages and index
values use two decimal places with RoundingMode.HALF_UP. This ensures
precise financial arithmetic across Quote, OrderBookEntry, MarketIndex,
and MarketDataSimulator.

+++ backend/src/main/java/com/bloomfield/terminal/model/Quote.java
@@ -1,16 +1,18 @@
 package com.bloomfield.terminal.model;
 
+import java.math.BigDecimal;
+
 public record Quote(
         String ticker,
         String name,
         String sector,
-        double price,
-        double open,
-        double high,
-        double low,
-        double close,
+        BigDecimal price,
+        BigDecimal open,
+        BigDecimal high,
+        BigDecimal low,
+        BigDecimal close,
         long volume,
-        double change,
-        double changePercent,
+        BigDecimal change,
+        BigDecimal changePercent,
         long timestamp
 ) {}
